pyproject.toml
[project]
name = "gv"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]==0.123.9",
    "uvicorn[standard]==0.38.0",
    "asyncpg==0.31.0",
    "sqlalchemy[asyncio]==2.0.44",
    "pydantic_settings==2.12.0",
    "secure==1.0.1"
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
venvPath = "."
venv = ".venv"

[dependency-groups]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-mock==3.15.1",
    "factory-boy==3.3.3",
    "httpx==0.28.1",
    "ruff==0.14.8",
    "pre-commit==4.5.0",
    "alembic==1.17.2",
    "psycopg==3.3.2",
    "psycopg-binary>=3.3.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]
